+++ app/assets/js/pages/CompanyWorkMapPage/index.tsx
@@ -1,16 +1,7 @@
-import * as React from "react";
 import { ShouldRevalidateFunction } from "react-router-dom";
 
-import * as Pages from "@/components/Pages";
-import { WorkMapItem, getWorkMap } from "@/models/workMap";
-import { Paths } from "@/routes/paths";
-import { redirectIfFeatureNotEnabled } from "@/routes/redirectIfFeatureEnabled";
-
-import { Page as PageContainer, WorkMap } from "turboui";
-
-interface LoaderResult {
-  workMap: WorkMapItem[];
-}
+export { Page } from "./page";
+export { loader } from "./loader";
 
 /**
  * Prevents the loader from rerunning when only the search parameters change.
@@ -23,31 +14,3 @@ export const shouldRevalidate: ShouldRevalidateFunction = ({ currentUrl, nextUrl
 
   return defaultShouldRevalidate;
 };
-
-export async function loader({ params }): Promise<LoaderResult> {
-  await redirectIfFeatureNotEnabled(params, {
-    feature: "work_map_page",
-    path: Paths.homePath(),
-  });
-
-  const workMap = await getWorkMap({}).then((data) => data.workMap || []);
-
-  return { workMap };
-}
-
-function useLoadedData(): LoaderResult {
-  return Pages.useLoadedData() as LoaderResult;
-}
-
-export function Page() {
-  const { workMap } = useLoadedData();
-  const title = "Company Work Map";
-
-  return (
-    <div className="py-6 px-2">
-      <PageContainer title={title} size="fullwidth">
-        <WorkMap title={title} items={workMap as WorkMap.Item[]} />
-      </PageContainer>
-    </div>
-  );
-}
