+++ app/knip.json
@@ -14,6 +14,7 @@
     "ee/assets/js/admin_api/index.tsx",
     "**/__tests__/**",
     "assets/js/models/companies/hasFeature.tsx",
+    "assets/js/pages/CompanyWorkMapPage/index.tsx"
   ],
   "ignoreDependencies": [
     "@types/jest",
